updated gui, and fixed bug where new training grids were overwritten within a set

+++ src/ocr/data/TrainingManager.java
@@ -17,7 +17,7 @@ public class TrainingManager {
         _trainingSet = trainingSet;
     }
 
-    public void setLearningRate(int learningRate) {
+    public void setLearningRate(double learningRate) {
         _learningRate = learningRate;
     }
